refactor: refactor codes
