--- packages/dart_flutter_common/lib/src/components/_export.dart
@@ -1,1 +1,0 @@
-export 'image.dart';
