+++ packages/dart_flutter_common/lib/src/string.dart
@@ -1,3 +1,4 @@
+/// [String] 型の拡張クラス。
 extension StringExtension on String {
   /// 空文字の場合に代わりの文字を返す。
   String ifIsEmpty(String placeholder) => isEmpty ? placeholder : this;
