+++ packages/dart_flutter_common/lib/src/widgets/_export.dart
@@ -1,0 +1,3 @@
+export 'generic_image.dart';
+export 'section.dart';
+export 'selectable_chips.dart';
